+++ src/Proto3/Suite/DotProto/Generate.hs
@@ -43,7 +43,7 @@ import           Data.Either                    (partitionEithers)
 import           Data.List                      (find, intercalate, nub, sortBy, stripPrefix)
 import qualified Data.List.NonEmpty             as NE
 import           Data.List.Split                (splitOn)
-import Data.List.NonEmpty (NonEmpty (..))
+import           Data.List.NonEmpty (NonEmpty (..))
 import qualified Data.Map                       as M
 import           Data.Maybe                     (fromMaybe)
 import           Data.Monoid
@@ -79,6 +79,7 @@ data CompileArgs = CompileArgs
   }
 
 data StringType = StringType String String
+  -- ^ Qualified module name, then unqualified type name.
 
 parseStringType :: String -> Either String StringType
 parseStringType str = case splitOn "." str of
@@ -391,11 +392,12 @@ msgTypeFromDpTypeInfo ctxt DotProtoTypeInfo{..} ident = do
     identName <- qualifiedMessageTypeName ctxt dotProtoTypeInfoParent ident
     pure $ HsTyCon (Qual modName (HsIdent identName))
 
-haskellName, jsonpbName, grpcName, protobufName, proxyName :: String -> HsQName
+haskellName, jsonpbName, grpcName, protobufName, protobufASTName, proxyName :: String -> HsQName
 haskellName  name = Qual (Module "Hs")         (HsIdent name)
 jsonpbName   name = Qual (Module "HsJSONPB")   (HsIdent name)
 grpcName     name = Qual (Module "HsGRPC")     (HsIdent name)
 protobufName name = Qual (Module "HsProtobuf") (HsIdent name)
+protobufASTName name = Qual (Module "HsProtobufAST") (HsIdent name)
 proxyName    name = Qual (Module "Proxy")      (HsIdent name)
 
 modulePathModName :: MonadError CompileError m => Path -> m Module
@@ -441,28 +443,44 @@ dhallInjectInstDecl typeName =
 
 -- ** Helpers to wrap/unwrap types for protobuf (de-)serialization
 
-coerceE :: Bool -> HsType -> HsType -> Maybe HsExp
-coerceE _ from to | from == to = Nothing
-coerceE unsafe from to = Just $ HsApp (HsApp coerceF (typeApp from)) (typeApp to)
+data FieldContext = WithinMessage | WithinOneOf
+  deriving (Eq, Show)
+
+typeApp :: HsType -> HsExp
+typeApp ty = uvar_ ("@("++ pp ty ++ ")")
   where
     -- Do not add linebreaks to typeapps as that causes parse errors
     pp = prettyPrintStyleMode style{mode=OneLineMode} defaultMode
-    typeApp ty = uvar_ ("@("++ pp ty ++ ")")
-    coerceF | unsafe = HsVar (haskellName "unsafeCoerce")
-            | otherwise  = HsVar (haskellName "coerce")
 
-wrapE :: MonadError CompileError m => StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> HsExp -> m HsExp
-wrapE stringType ctxt opts dpt e = maybe e (\f -> apply f [e]) <$>
-  (coerceE (isMap dpt) <$> dptToHsType stringType ctxt dpt <*> dptToHsTypeWrapped stringType opts ctxt dpt)
-
-unwrapE :: MonadError CompileError m => StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> HsExp -> m HsExp
-unwrapE stringType ctxt opts dpt e = maybe e (\f -> apply f [e]) <$>
-   (coerceE (isMap dpt) <$>
-     overParser (dptToHsTypeWrapped stringType opts ctxt dpt) <*>
-       overParser (dptToHsType stringType ctxt dpt))
+coerceE :: Bool -> Bool -> HsType -> HsType -> Maybe HsExp
+coerceE _ _ from to | from == to = Nothing
+coerceE overTyCon unsafe from to =
+    Just $ HsApp (HsApp coerceF (typeApp from)) (typeApp to)
   where
-    overParser = fmap $ HsTyApp (HsTyVar (HsIdent "_"))
+    coerceF | unsafe = HsVar (name "unsafeCoerce")
+            | otherwise  = HsVar (name "coerce")
+    name | overTyCon = protobufName . (<> "Over")
+         | otherwise = haskellName
 
+wrapFunE :: MonadError CompileError m => Bool -> FieldContext -> StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> m (Maybe HsExp)
+wrapFunE overTyCon fc stringType ctxt opts dpt =
+  coerceE overTyCon (isMap dpt)
+    <$> dptToHsType fc stringType ctxt dpt
+    <*> dptToHsTypeWrapped fc stringType opts ctxt dpt
+
+wrapE :: MonadError CompileError m => FieldContext -> StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> HsExp -> m HsExp
+wrapE fc stringType ctxt opts dpt e =
+  flip applyMaybe e <$> wrapFunE False fc stringType ctxt opts dpt
+
+unwrapFunE :: MonadError CompileError m => Bool -> FieldContext -> StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> m (Maybe HsExp)
+unwrapFunE overTyCon fc stringType ctxt opts dpt =
+  coerceE overTyCon (isMap dpt)
+    <$> dptToHsTypeWrapped fc stringType opts ctxt dpt
+    <*> dptToHsType fc stringType ctxt dpt
+
+unwrapE :: MonadError CompileError m => FieldContext -> StringType -> TypeContext -> [DotProtoOption] -> DotProtoType -> HsExp -> m HsExp
+unwrapE fc stringType ctxt opts dpt e = do
+  flip applyMaybe e <$> unwrapFunE True fc stringType ctxt opts dpt
 
 --------------------------------------------------------------------------------
 --
@@ -470,61 +488,90 @@ unwrapE stringType ctxt opts dpt e = maybe e (\f -> apply f [e]) <$>
 --
 
 -- | Convert a dot proto type to a Haskell type
-dptToHsType :: MonadError CompileError m => StringType -> TypeContext -> DotProtoType -> m HsType
-dptToHsType = foldDPT dptToHsContType . dpptToHsType
+dptToHsType :: MonadError CompileError m => FieldContext -> StringType -> TypeContext -> DotProtoType -> m HsType
+dptToHsType fc = foldDPT (dptToHsContType fc) . dpptToHsType
 
 -- | Convert a dot proto type to a wrapped Haskell type
 dptToHsTypeWrapped
   :: MonadError CompileError m
-  => StringType
+  => FieldContext
+  -> StringType
   -> [DotProtoOption]
   -> TypeContext
   -> DotProtoType
   -> m HsType
-dptToHsTypeWrapped (StringType _ stringType) opts =
-   foldDPT
-     -- The wrapper for the collection type replaces the native haskell
-     -- collection type, so try that first.
-     (\ctxt ty -> maybe (dptToHsContType ctxt ty) id (dptToHsWrappedContType ctxt opts ty))
-     -- Always wrap the primitive type.
-     (\ctxt ty -> dpptToHsTypeWrapper ty <$> dpptToHsType' ctxt ty)
-  where
-    dpptToHsType' :: MonadError CompileError m
-                  => TypeContext
-                  -> DotProtoPrimType
-                  -> m HsType
-    dpptToHsType' ctxt =  \case
-      Int32    -> pure $ primType_ "Int32"
-      Int64    -> pure $ primType_ "Int64"
-      SInt32   -> pure $ primType_ "Int32"
-      SInt64   -> pure $ primType_ "Int64"
-      UInt32   -> pure $ primType_ "Word32"
-      UInt64   -> pure $ primType_ "Word64"
-      Fixed32  -> pure $ primType_ "Word32"
-      Fixed64  -> pure $ primType_ "Word64"
-      SFixed32 -> pure $ primType_ "Int32"
-      SFixed64 -> pure $ primType_ "Int64"
-      String   -> pure $ primType_ stringType
-      Bytes    -> pure $ primType_ "ByteString"
-      Bool     -> pure $ primType_ "Bool"
-      Float    -> pure $ primType_ "Float"
-      Double   -> pure $ primType_ "Double"
-      Named (Dots (Path ("google" :| ["protobuf", x])))
-        | x == "Int32Value" -> pure $ protobufWrapperType_ "Int32"
-        | x == "Int64Value" -> pure $ protobufWrapperType_ "Int64"
-        | x == "UInt32Value" -> pure $ protobufWrapperType_ "Word32"
-        | x == "UInt64Value" -> pure $ protobufWrapperType_ "Word64"
-        | x == "StringValue" -> pure $ protobufWrapperType_ "Text"
-        | x == "BytesValue" -> pure $ protobufWrapperType_ "ByteString"
-        | x == "BoolValue" -> pure $ protobufWrapperType_ "Bool"
-        | x == "FloatValue" -> pure $ protobufWrapperType_ "Float"
-        | x == "DoubleValue" -> pure $ protobufWrapperType_ "Double"
-      Named msgName ->
-        case M.lookup msgName ctxt of
-          Just ty@(DotProtoTypeInfo { dotProtoTypeInfoKind = DotProtoKindEnum }) ->
-              HsTyApp (protobufType_ "Enumerated") <$> msgTypeFromDpTypeInfo ctxt ty msgName
-          Just ty -> msgTypeFromDpTypeInfo ctxt ty msgName
-          Nothing -> noSuchTypeError msgName
+dptToHsTypeWrapped fc stringType opts =
+  foldDPT
+    -- The wrapper for the collection type replaces the native haskell
+    -- collection type, so try that first.
+    (\ctxt ty -> maybe (dptToHsContType fc ctxt ty) id (dptToHsWrappedContType fc ctxt opts ty))
+    -- Always wrap the primitive type.
+    (dpptToHsTypeWrapped stringType)
+
+-- | Like 'dptToHsTypeWrapped' but without use of
+-- 'dptToHsContType' or 'dptToHsWrappedContType'.
+dpptToHsTypeWrapped
+  :: MonadError CompileError m
+  => StringType
+  -> TypeContext
+  -> DotProtoPrimType
+  -> m HsType
+dpptToHsTypeWrapped (StringType _ stringType) ctxt =  \case
+  Int32 ->
+    pure $ primType_ "Int32"
+  Int64 ->
+    pure $ primType_ "Int64"
+  SInt32 ->
+    pure $ protobufSignedType_ $ primType_ "Int32"
+  SInt64 ->
+    pure $ protobufSignedType_ $ primType_ "Int64"
+  UInt32 ->
+    pure $ primType_ "Word32"
+  UInt64 ->
+    pure $ primType_ "Word64"
+  Fixed32 ->
+    pure $ protobufFixedType_ $ primType_ "Word32"
+  Fixed64 ->
+    pure $ protobufFixedType_ $ primType_ "Word64"
+  SFixed32 ->
+    pure $ protobufSignedType_ $ protobufFixedType_ $ primType_ "Int32"
+  SFixed64 ->
+    pure $ protobufSignedType_ $ protobufFixedType_ $ primType_ "Int64"
+  String ->
+    pure $ protobufStringType_ stringType
+  Bytes  ->
+    pure $ protobufBytesType_ "ByteString"
+  Bool ->
+    pure $ primType_ "Bool"
+  Float ->
+    pure $ primType_ "Float"
+  Double ->
+    pure $ primType_ "Double"
+  Named (Dots (Path ("google" :| ["protobuf", x])))
+    | x == "Int32Value" ->
+        pure $ protobufWrappedType_ $ primType_ "Int32"
+    | x == "Int64Value" ->
+        pure $ protobufWrappedType_ $ primType_ "Int64"
+    | x == "UInt32Value" ->
+        pure $ protobufWrappedType_ $ primType_ "Word32"
+    | x == "UInt64Value" ->
+        pure $ protobufWrappedType_ $ primType_ "Word64"
+    | x == "StringValue" ->
+        pure $ protobufWrappedType_ $ protobufStringType_ stringType
+    | x == "BytesValue" ->
+        pure $ protobufWrappedType_ $ protobufBytesType_ "ByteString"
+    | x == "BoolValue" ->
+        pure $ protobufWrappedType_ $ primType_ "Bool"
+    | x == "FloatValue" ->
+        pure $ protobufWrappedType_ $ primType_ "Float"
+    | x == "DoubleValue" ->
+        pure $ protobufWrappedType_ $ primType_ "Double"
+  Named msgName ->
+    case M.lookup msgName ctxt of
+      Just ty@(DotProtoTypeInfo { dotProtoTypeInfoKind = DotProtoKindEnum }) ->
+          HsTyApp (protobufType_ "Enumerated") <$> msgTypeFromDpTypeInfo ctxt ty msgName
+      Just ty -> msgTypeFromDpTypeInfo ctxt ty msgName
+      Nothing -> noSuchTypeError msgName
 
 foldDPT :: MonadError CompileError m
         => (TypeContext -> DotProtoType -> HsType -> HsType)
@@ -547,10 +594,14 @@ foldDPT dptToHsCont foldPrim ctxt dpt =
 
 -- | Translate DotProtoType constructors to wrapped Haskell container types
 -- (for Message serde instances).
-dptToHsWrappedContType :: TypeContext -> [DotProtoOption] -> DotProtoType -> Maybe (HsType -> HsType)
-dptToHsWrappedContType ctxt opts = \case
+--
+-- When the given 'FieldContext' is 'WithinOneOf' we do not wrap submessages
+-- in "Maybe" because the entire oneof is already wrapped in a "Maybe".
+dptToHsWrappedContType :: FieldContext -> TypeContext -> [DotProtoOption] -> DotProtoType -> Maybe (HsType -> HsType)
+dptToHsWrappedContType fc ctxt opts = \case
   Prim (Named tyName)
-    | isMessage ctxt tyName -> Just $ HsTyApp (protobufType_ "Nested")
+    | WithinMessage <- fc, isMessage ctxt tyName
+                            -> Just $ HsTyApp (protobufType_ "Nested")
   Repeated (Named tyName)
     | isMessage ctxt tyName -> Just $ HsTyApp (protobufType_ "NestedVec")
   Repeated ty
@@ -561,25 +612,17 @@ dptToHsWrappedContType ctxt opts = \case
   _ -> Nothing
 
 -- | Translate DotProtoType to Haskell container types.
-dptToHsContType :: TypeContext -> DotProtoType -> HsType -> HsType
-dptToHsContType ctxt = \case
-  Prim (Named tyName) | isMessage ctxt tyName
+--
+-- When the given 'FieldContext' is 'WithinOneOf' we do not wrap submessages
+-- in "Maybe" because the entire oneof is already wrapped in a "Maybe".
+dptToHsContType :: FieldContext -> TypeContext -> DotProtoType -> HsType -> HsType
+dptToHsContType fc ctxt = \case
+  Prim (Named tyName) | WithinMessage <- fc, isMessage ctxt tyName
                      -> HsTyApp $ primType_ "Maybe"
   Repeated _         -> HsTyApp $ primType_ "Vector"
   NestedRepeated _   -> HsTyApp $ primType_ "Vector"
   Map _ _            -> HsTyApp $ primType_ "Map"
   _                  -> id
-
--- | Haskell wrapper for primitive dot proto types
-dpptToHsTypeWrapper :: DotProtoPrimType -> HsType -> HsType
-dpptToHsTypeWrapper = \case
-  SInt32   -> HsTyApp (protobufType_ "Signed")
-  SInt64   -> HsTyApp (protobufType_ "Signed")
-  SFixed32 -> HsTyApp (protobufType_ "Signed") . HsTyApp (protobufType_ "Fixed")
-  SFixed64 -> HsTyApp (protobufType_ "Signed") . HsTyApp (protobufType_ "Fixed")
-  Fixed32  -> HsTyApp (protobufType_ "Fixed")
-  Fixed64  -> HsTyApp (protobufType_ "Fixed")
-  _        -> id
 
 -- | Convert a dot proto prim type to an unwrapped Haskell type
 dpptToHsType :: MonadError CompileError m
@@ -608,7 +651,7 @@ dpptToHsType (StringType _ stringType) ctxt = \case
     | x == "Int64Value" -> pure $ primType_ "Int64"
     | x == "UInt32Value" -> pure $ primType_ "Word32"
     | x == "UInt64Value" -> pure $ primType_ "Word64"
-    | x == "StringValue" -> pure $ primType_ "Text"
+    | x == "StringValue" -> pure $ primType_ stringType
     | x == "BytesValue" -> pure $ primType_ "ByteString"
     | x == "BoolValue" -> pure $ primType_ "Bool"
     | x == "FloatValue" -> pure $ primType_ "Float"
@@ -686,9 +729,9 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
             defaultMessageDeriving
 
     let getName = \case
-          DotProtoMessageField fld     -> [dotProtoFieldName fld]
-          DotProtoMessageOneOf ident _ -> [ident]
-          _                            -> []
+          DotProtoMessageField fld -> (: []) <$> getFieldNameForSchemaInstanceDeclaration fld
+          DotProtoMessageOneOf ident _ -> (: []) . (Nothing, ) <$> dpIdentUnqualName ident
+          _ -> pure []
 
     foldMapM id
       [ sequence
@@ -697,8 +740,8 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
           , pure (hasDefaultInstD messageName)
           , messageInstD stringType ctxt' parentIdent messageIdent messageParts
 
-          , toJSONPBMessageInstD   ctxt' parentIdent messageIdent messageParts
-          , fromJSONPBMessageInstD ctxt' parentIdent messageIdent messageParts
+          , toJSONPBMessageInstD stringType ctxt' parentIdent messageIdent messageParts
+          , fromJSONPBMessageInstD stringType ctxt' parentIdent messageIdent messageParts
 
             -- Generate Aeson instances in terms of JSONPB instances
           , pure (toJSONInstDecl messageName)
@@ -706,8 +749,8 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
 
 #ifdef SWAGGER
           -- And the Swagger ToSchema instance corresponding to JSONPB encodings
-          , toSchemaInstanceDeclaration messageName Nothing
-              =<< foldMapM (traverse dpIdentUnqualName . getName) messageParts
+          , toSchemaInstanceDeclaration stringType ctxt' messageName Nothing
+              =<< foldMapM getName messageParts
 #endif
 
 #ifdef DHALL
@@ -735,7 +778,7 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
     messagePartFieldD :: String -> DotProtoMessagePart -> m [([HsName], HsBangType)]
     messagePartFieldD messageName (DotProtoMessageField DotProtoField{..}) = do
       fullName <- prefixedFieldName messageName =<< dpIdentUnqualName dotProtoFieldName
-      fullTy <- dptToHsType stringType ctxt' dotProtoFieldType
+      fullTy <- dptToHsType WithinMessage stringType ctxt' dotProtoFieldType
       pure [ ([HsIdent fullName], HsUnBangedTy fullTy ) ]
 
     messagePartFieldD messageName (DotProtoMessageOneOf fieldName _) = do
@@ -764,8 +807,8 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
       (cons, idents) <- fmap unzip (mapM (oneOfCons fullName) fields)
 
 #ifdef SWAGGER
-      toSchemaInstance <- toSchemaInstanceDeclaration fullName (Just idents)
-                            =<< mapM (dpIdentUnqualName . dotProtoFieldName) fields
+      toSchemaInstance <- toSchemaInstanceDeclaration stringType ctxt' fullName (Just idents)
+                            =<< mapM getFieldNameForSchemaInstanceDeclaration fields
 #endif
 
       pure [ dataDecl_ fullName cons defaultMessageDeriving
@@ -782,12 +825,7 @@ dotProtoMessageD stringType ctxt parentIdent messageIdent messageParts = do
 
     oneOfCons :: String -> DotProtoField -> m (HsConDecl, HsName)
     oneOfCons fullName DotProtoField{..} = do
-       consTy <- case dotProtoFieldType of
-            Prim msg@(Named msgName)
-              | isMessage ctxt' msgName
-                -> dpptToHsType stringType ctxt' msg
-            _   -> dptToHsType stringType ctxt' dotProtoFieldType
-
+       consTy <- dptToHsType WithinOneOf stringType ctxt' dotProtoFieldType
        consName <- prefixedConName fullName =<< dpIdentUnqualName dotProtoFieldName
        let ident = HsIdent consName
        pure (conDecl_ ident [HsUnBangedTy consTy], ident)
@@ -855,7 +893,7 @@ messageInstD stringType ctxt parentIdent msgIdent messageParts = do
       let recordFieldName' = uvar_ (coerce recordFieldName) in
       case fieldInfo of
         FieldNormal _fieldName fieldNum dpType options -> do
-            fieldE <- wrapE stringType ctxt options dpType recordFieldName'
+            fieldE <- wrapE WithinMessage stringType ctxt options dpType recordFieldName'
             pure $ apply encodeMessageFieldE [ fieldNumberE fieldNum, fieldE ]
 
         FieldOneOf OneofField{subfields} -> do
@@ -882,7 +920,11 @@ messageInstD stringType ctxt parentIdent msgIdent messageParts = do
               let wrapJust = HsParen . HsApp (HsVar (haskellName "Just"))
 
               xE <- (if isMaybe then id else fmap forceEmitE)
-                     . wrapE stringType ctxt options dpType
+                     . wrapE WithinMessage stringType ctxt options dpType
+                         -- For now we use 'WithinMessage' to preserve
+                         -- the historical approach of treating this field
+                         -- as if it were an ordinary non-oneof field that
+                         -- just happens to be present.
                      . (if isMaybe then wrapJust else id)
                      $ uvar_ "y"
 
@@ -895,7 +937,8 @@ messageInstD stringType ctxt parentIdent msgIdent messageParts = do
     decodeMessageField QualifiedField{fieldInfo} =
       case fieldInfo of
         FieldNormal _fieldName fieldNum dpType options ->
-            unwrapE stringType ctxt options dpType $ apply atE [ decodeMessageFieldE, fieldNumberE fieldNum ]
+            unwrapE WithinMessage stringType ctxt options dpType $
+              apply atE [ decodeMessageFieldE, fieldNumberE fieldNum ]
 
         FieldOneOf OneofField{subfields} -> do
             parsers <- mapM subfieldParserE subfields
@@ -912,7 +955,11 @@ messageInstD stringType ctxt parentIdent msgIdent messageParts = do
                                            composeOp
                                            (uvar_ consName))
 
-              alts <- unwrapE stringType ctxt options dpType decodeMessageFieldE
+              -- For now we continue the historical practice of parsing
+              -- submessages within oneofs as if were outside of oneofs,
+              -- and replacing the "Just . Ctor" with "fmap . Ctor".
+              -- That is why we do not pass WithinOneOf.
+              alts <- unwrapE WithinMessage stringType ctxt options dpType decodeMessageFieldE
 
               pure $ HsTuple
                    [ fieldNumberE fieldNumber
@@ -923,48 +970,61 @@ messageInstD stringType ctxt parentIdent msgIdent messageParts = do
 -- *** Generate ToJSONPB/FromJSONPB instances
 
 toJSONPBMessageInstD
-    :: MonadError CompileError m
-    => TypeContext
+    :: forall m
+     . MonadError CompileError m
+    => StringType
+    -> TypeContext
     -> DotProtoIdentifier
     -> DotProtoIdentifier
     -> [DotProtoMessagePart]
     -> m HsDecl
-toJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
+toJSONPBMessageInstD stringType ctxt parentIdent msgIdent messageParts = do
     msgName    <- qualifiedMessageName parentIdent msgIdent
     qualFields <- getQualifiedFields msgName messageParts
 
-    let applyE nm oneofNm =
-          apply (HsVar (jsonpbName nm))
-                [ HsList (foldQF defPairE (oneofCaseE oneofNm) <$> qualFields) ]
+    let applyE nm oneofNm = do
+          fs <- traverse (encodeMessageField oneofNm) qualFields
+          pure $ apply (HsVar (jsonpbName nm)) [HsList fs]
 
     let patBinder = foldQF (const fieldBinder) (oneofSubDisjunctBinder . subfields)
-    let matchE nm appNm oneofAppNm =
-          match_
+    let matchE nm appNm oneofAppNm = do
+          rhs <- applyE appNm oneofAppNm
+          pure $ match_
             (HsIdent nm)
             [ HsPApp (unqual_ msgName)
                      (patVar . patBinder <$> qualFields) ]
-            (HsUnGuardedRhs (applyE appNm oneofAppNm))
+            (HsUnGuardedRhs rhs)
             []
+
+    toJSONPB <- matchE "toJSONPB" "object" "objectOrNull"
+    toEncoding <- matchE "toEncodingPB" "pairs" "pairsOrNull"
 
     pure $ instDecl_ (jsonpbName "ToJSONPB")
                      [ type_ msgName ]
-                     [ HsFunBind [matchE "toJSONPB"     "object" "objectOrNull"]
-                     , HsFunBind [matchE "toEncodingPB" "pairs"  "pairsOrNull" ]
+                     [ HsFunBind [toJSONPB]
+                     , HsFunBind [toEncoding]
                      ]
 
   where
+    encodeMessageField :: String -> QualifiedField -> m HsExp
+    encodeMessageField oneofNm (QualifiedField _ fieldInfo) =
+      case fieldInfo of
+        FieldNormal fldName fldNum dpType options ->
+          defPairE fldName fldNum dpType options
+        FieldOneOf oo ->
+          oneofCaseE oneofNm oo
+
     -- E.g.
     -- "another" .= f2 -- always succeeds (produces default value on missing field)
-    defPairE fldName fldNum =
-      HsInfixApp (str_ (coerce fldName))
-                 toJSONPBOp
-                 (uvar_ (fieldBinder fldNum))
+    defPairE fldName fldNum dpType options = do
+      w <- wrapE WithinMessage stringType ctxt options dpType (uvar_ (fieldBinder fldNum))
+      pure $ HsInfixApp (str_ (coerce fldName)) toJSONPBOp w
 
     -- E.g.
     -- HsJSONPB.pair "name" f4 -- fails on missing field
-    pairE fldNm varNm =
-      apply (HsVar (jsonpbName "pair"))
-            [ str_ (coerce fldNm) , uvar_ varNm]
+    oneOfPairE fldNm varNm options dpType = do
+      w <- wrapE WithinOneOf stringType ctxt options dpType (uvar_ varNm)
+      pure $ apply (HsVar (jsonpbName "pair")) [str_ (coerce fldNm), w]
 
     -- Suppose we have a sum type Foo, nested inside a message Bar.
     -- We want to generate the following:
@@ -979,9 +1039,11 @@ toJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
     -- >     , <encode more>
     -- >     , <encode stuff>
     -- >     ]
-    oneofCaseE retJsonCtor (OneofField typeName subfields) =
-        HsParen
-          $ HsLet [ HsFunBind [ match_ (HsIdent caseName) [] (HsUnGuardedRhs caseExpr) [] ] ]
+    oneofCaseE :: String -> OneofField -> m HsExp
+    oneofCaseE retJsonCtor (OneofField typeName subfields) = do
+        altEs <- traverse altE subfields
+        pure $ HsParen
+          $ HsLet [ HsFunBind [ match_ (HsIdent caseName) [] (HsUnGuardedRhs (caseExpr altEs)) [] ] ]
           $ HsLambda defaultSrcLoc [patVar optsStr] (HsIf dontInline noInline yesInline)
       where
         optsStr = "options"
@@ -1000,6 +1062,16 @@ toJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
 
         yesInline = HsApp caseBnd opts
 
+        altE sub@(OneofSubfield _ conName pbFldNm dpType options) = do
+          let patVarNm = oneofSubBinder sub
+          p <- oneOfPairE pbFldNm patVarNm options dpType
+          pure $ alt_ (HsPApp (haskellName "Just")
+                              [ HsPParen
+                                (HsPApp (unqual_ conName) [patVar patVarNm])
+                              ]
+                      )
+                      (HsUnGuardedAlt p)
+                      []
 
         -- E.g.
         -- case f4_or_f9 of
@@ -1009,35 +1081,29 @@ toJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
         --     -> HsJSONPB.pair "someid" f9
         --   Nothing
         --     -> mempty
-        caseExpr = HsParen $
+        caseExpr altEs = HsParen $
             HsCase disjunctName (altEs <> [fallthroughE])
           where
             disjunctName = uvar_ (oneofSubDisjunctBinder subfields)
-            altEs = do
-              sub@(OneofSubfield _ conName pbFldNm _ _) <- subfields
-              let patVarNm = oneofSubBinder sub
-              pure $ alt_ (HsPApp (haskellName "Just")
-                                  [ HsPParen
-                                    (HsPApp (unqual_ conName) [patVar patVarNm])
-                                  ]
-                          )
-                          (HsUnGuardedAlt (pairE pbFldNm patVarNm))
-                          []
             fallthroughE =
               alt_ (HsPApp (haskellName "Nothing") [])
                    (HsUnGuardedAlt memptyE)
                    []
 
 fromJSONPBMessageInstD
-    :: MonadError CompileError m
-    => TypeContext
+    :: forall m
+     . MonadError CompileError m
+    => StringType
+    -> TypeContext
     -> DotProtoIdentifier
     -> DotProtoIdentifier
     -> [DotProtoMessagePart]
     -> m HsDecl
-fromJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
+fromJSONPBMessageInstD stringType ctxt parentIdent msgIdent messageParts = do
     msgName    <- qualifiedMessageName parentIdent msgIdent
     qualFields <- getQualifiedFields msgName messageParts
+
+    fieldParsers <- traverse parseField qualFields
 
     let parseJSONPBE =
           apply (HsVar (jsonpbName "withObject"))
@@ -1047,7 +1113,7 @@ fromJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
           where
             fieldAps = foldl (\f -> HsInfixApp f apOp)
                              (apply pureE [ uvar_ msgName ])
-                             (foldQF normalParserE oneofParserE <$> qualFields)
+                             fieldParsers
 
     let parseJSONPBDecl =
           match_ (HsIdent "parseJSONPB") [] (HsUnGuardedRhs parseJSONPBE) []
@@ -1058,6 +1124,11 @@ fromJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
   where
     lambdaPVar = patVar "obj"
     lambdaVar  = uvar_ "obj"
+
+    parseField (QualifiedField _ (FieldNormal fldName _ dpType options)) =
+      normalParserE fldName dpType options
+    parseField (QualifiedField _ (FieldOneOf fld)) =
+      oneofParserE fld
 
     -- E.g., for message
     --   message Something { oneof name_or_id { string name = _; int32 someid = _; } }
@@ -1070,10 +1141,12 @@ fromJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
     --     <|>
     --     (parseSomethingNameOrId obj)
     -- )
-    oneofParserE (OneofField oneofType fields) =
-        HsParen $
+    oneofParserE :: OneofField -> m HsExp
+    oneofParserE (OneofField oneofType fields) = do
+        ds <- tryParseDisjunctsE
+        pure $ HsParen $
           HsLet [ HsFunBind [ match_ (HsIdent letBndStr) [patVar letArgStr ]
-                                     (HsUnGuardedRhs tryParseDisjunctsE) []
+                                     (HsUnGuardedRhs ds) []
                             ]
                 ]
                 (HsInfixApp parseWrapped altOp parseUnwrapped)
@@ -1098,26 +1171,33 @@ fromJSONPBMessageInstD _ctxt parentIdent msgIdent messageParts = do
         --     , (Just . SomethingPickOneSomeid) <$> (HsJSONPB.parseField parseObj "someid")
         --     , pure Nothing
         --     ]
-        tryParseDisjunctsE =
-            HsApp msumE (HsList (map subParserE fields <> fallThruE))
-          where
-            fallThruE
-              = [ HsApp pureE (HsVar (haskellName "Nothing")) ]
-            subParserE OneofSubfield{subfieldConsName, subfieldName}
-              = HsInfixApp
-                  (HsInfixApp (HsVar (haskellName "Just"))
-                              composeOp
-                              (uvar_ subfieldConsName))
-                  fmapOp
-                  (apply (HsVar (jsonpbName "parseField"))
-                         [ letArgName
-                         , str_ (coerce subfieldName)])
+        tryParseDisjunctsE = do
+          fs <- traverse subParserE fields
+          pure $ HsApp msumE (HsList (fs <> fallThruE))
+
+        fallThruE = [ HsApp pureE (HsVar (haskellName "Nothing")) ]
+
+        subParserE OneofSubfield{subfieldConsName, subfieldName,
+                                 subfieldType, subfieldOptions} = do
+          maybeCoercion <-
+            unwrapFunE False WithinOneOf stringType ctxt subfieldOptions subfieldType
+          let inject = (HsInfixApp (HsVar (haskellName "Just"))
+                                   composeOp
+                                   (uvar_ subfieldConsName))
+          pure $ HsInfixApp
+              (maybe inject (HsInfixApp inject composeOp) maybeCoercion)
+              fmapOp
+              (apply (HsVar (jsonpbName "parseField"))
+                     [ letArgName
+                     , str_ (coerce subfieldName)])
 
     -- E.g. obj .: "someid"
-    normalParserE fldNm _ =
-      HsInfixApp lambdaVar
-                 parseJSONPBOp
-                 (str_(coerce fldNm))
+    normalParserE :: FieldName -> DotProtoType -> [DotProtoOption] -> m HsExp
+    normalParserE fldName dpType options =
+      unwrapE WithinMessage stringType ctxt options dpType $
+        HsInfixApp lambdaVar
+                   parseJSONPBOp
+                   (str_(coerce fldName))
 
 -- *** Generate default Aeson To/FromJSON and Swagger ToSchema instances
 -- (These are defined in terms of ToJSONPB)
@@ -1146,16 +1226,30 @@ fromJSONInstDecl typeName =
 
 -- *** Generate `ToSchema` instance
 
+getFieldNameForSchemaInstanceDeclaration
+  :: MonadError CompileError m
+  => DotProtoField
+  -> m (Maybe ([DotProtoOption], DotProtoType), String)
+getFieldNameForSchemaInstanceDeclaration fld = do
+  unqual <- dpIdentUnqualName (dotProtoFieldName fld)
+  let optsType = (dotProtoFieldOptions fld, dotProtoFieldType fld)
+  pure (Just optsType, unqual)
+
 toSchemaInstanceDeclaration
     :: MonadError CompileError m
-    => String
+    => StringType
+    -> TypeContext
+    -> String
     -- ^ Name of the message type to create an instance for
     -> Maybe [HsName]
     -- ^ Oneof constructors
-    -> [String]
-    -- ^ Field names
+    -> [(Maybe ([DotProtoOption], DotProtoType), String)]
+    -- ^ Field names, with every field that is not actually a oneof
+    -- combining fields paired with its options and protobuf type
     -> m HsDecl
-toSchemaInstanceDeclaration messageName maybeConstructors fieldNames = do
+toSchemaInstanceDeclaration stringType ctxt messageName maybeConstructors fieldNamesEtc = do
+  let fieldNames = map snd fieldNamesEtc
+
   qualifiedFieldNames <- mapM (prefixedFieldName messageName) fieldNames
 
   let messageConstructor = HsCon (UnQual (HsIdent messageName))
@@ -1237,10 +1331,11 @@ toSchemaInstanceDeclaration messageName maybeConstructors fieldNames = do
 
   let toDeclareName fieldName = "declare_" ++ fieldName
 
-  let toArgument fieldName = HsApp asProxy declare
+  let toArgument fc (maybeOptsType, fieldName) =
+          maybe pure (uncurry (unwrapE fc stringType ctxt)) maybeOptsType $
+            HsApp asProxy declare
         where
           declare = uvar_ (toDeclareName fieldName)
-
           asProxy = HsVar (jsonpbName "asProxy")
 
       -- do let declare_fieldName0 = HsJSONPB.declareSchemaRef
@@ -1250,35 +1345,32 @@ toSchemaInstanceDeclaration messageName maybeConstructors fieldNames = do
       --    ...
       --    let _ = pure MessageName <*> HsJSONPB.asProxy declare_fieldName0 <*> HsJSONPB.asProxy declare_fieldName1 <*> ...
       --    return (...)
-  let expressionForMessage =
-          HsDo (bindingStatements ++ inferenceStatement ++ [ returnStatement ])
-        where
-          bindingStatements = do
-            (fieldName, qualifiedFieldName) <- zip fieldNames qualifiedFieldNames
+  let expressionForMessage = do
+        let bindingStatements = do
+              (fieldName, qualifiedFieldName) <- zip fieldNames qualifiedFieldNames
 
-            let declareIdentifier = HsIdent (toDeclareName fieldName)
+              let declareIdentifier = HsIdent (toDeclareName fieldName)
 
-            let stmt0 = HsLetStmt [ HsFunBind
-                                    [ HsMatch defaultSrcLoc declareIdentifier []
-                                               (HsUnGuardedRhs (HsVar (jsonpbName "declareSchemaRef"))) []
+              let stmt0 = HsLetStmt [ HsFunBind
+                                      [ HsMatch defaultSrcLoc declareIdentifier []
+                                                 (HsUnGuardedRhs (HsVar (jsonpbName "declareSchemaRef"))) []
+                                      ]
                                     ]
-                                  ]
 
-            let stmt1 = HsGenerator defaultSrcLoc (HsPVar (HsIdent qualifiedFieldName))
-                                      (HsApp (HsVar (UnQual declareIdentifier))
-                                             (HsCon (proxyName "Proxy")))
-            [ stmt0, stmt1]
+              let stmt1 = HsGenerator defaultSrcLoc (HsPVar (HsIdent qualifiedFieldName))
+                                        (HsApp (HsVar (UnQual declareIdentifier))
+                                               (HsCon (proxyName "Proxy")))
+              [ stmt0, stmt1]
 
+        inferenceStatement <- do
+          arguments <- traverse (toArgument WithinMessage) fieldNamesEtc
+          let patternBind = HsPatBind defaultSrcLoc HsPWildCard
+                (HsUnGuardedRhs (applicativeApply messageConstructor arguments)) []
+          pure $ if null fieldNames then [] else [ HsLetStmt [ patternBind ] ]
 
-          inferenceStatement =
-              if null fieldNames then [] else [ HsLetStmt [ patternBind ] ]
-            where
-              arguments = map toArgument fieldNames
+        let returnStatement = HsQualifier (HsApp returnE (HsParen namedSchema))
 
-              patternBind = HsPatBind defaultSrcLoc HsPWildCard
-                                        (HsUnGuardedRhs (applicativeApply messageConstructor arguments)) []
-
-          returnStatement = HsQualifier (HsApp returnE (HsParen namedSchema))
+        pure $ HsDo (bindingStatements ++ inferenceStatement ++ [ returnStatement ])
 
       -- do let declare_fieldName0 = HsJSONPB.declareSchemaRef
       --    let _ = pure ConstructorName0 <*> HsJSONPB.asProxy declare_fieldName0
@@ -1288,35 +1380,36 @@ toSchemaInstanceDeclaration messageName maybeConstructors fieldNames = do
       --    qualifiedFieldName1 <- declare_fieldName1 Proxy.Proxy
       --    ...
       --    return (...)
-  let expressionForOneOf constructors =
-          HsDo (bindingStatements ++ [ returnStatement ])
-        where
-          bindingStatements = do
-            (fieldName, qualifiedFieldName, constructor)
-                <- zip3 fieldNames qualifiedFieldNames constructors
+  let expressionForOneOf constructors = do
+        let bindingStatement (fieldNameEtc, qualifiedFieldName, constructor) = do
+              let declareIdentifier = HsIdent (toDeclareName (snd fieldNameEtc))
 
-            let declareIdentifier = HsIdent (toDeclareName fieldName)
+              let stmt0 = HsLetStmt [ HsFunBind
+                                        [ HsMatch defaultSrcLoc declareIdentifier []
+                                                   (HsUnGuardedRhs (HsVar (jsonpbName "declareSchemaRef"))) []
+                                        ]
+                                    ]
+              let stmt1 = HsGenerator defaultSrcLoc (HsPVar (HsIdent qualifiedFieldName))
+                                        (HsApp (HsVar (UnQual declareIdentifier))
+                                               (HsCon (proxyName "Proxy")))
+              inferenceStatement <- do
+                argument <- toArgument WithinOneOf fieldNameEtc
+                let patternBind = HsPatBind defaultSrcLoc HsPWildCard
+                      (HsUnGuardedRhs (applicativeApply (HsCon (UnQual constructor)) [ argument ])) []
+                pure $ if null fieldNames then [] else [ HsLetStmt [ patternBind ] ]
 
-            let stmt0 = HsLetStmt [ HsFunBind
-                                      [ HsMatch defaultSrcLoc declareIdentifier []
-                                                 (HsUnGuardedRhs (HsVar (jsonpbName "declareSchemaRef"))) []
-                                      ]
-                                  ]
-            let stmt1 = HsGenerator defaultSrcLoc (HsPVar (HsIdent qualifiedFieldName))
-                                      (HsApp (HsVar (UnQual declareIdentifier))
-                                             (HsCon (proxyName "Proxy")))
-            let inferenceStatement =
-                    if null fieldNames then [] else [ HsLetStmt [ patternBind ] ]
-                  where
-                    arguments = [ toArgument fieldName ]
+              pure $ [stmt0, stmt1] ++ inferenceStatement
 
-                    patternBind = HsPatBind defaultSrcLoc HsPWildCard
-                                              (HsUnGuardedRhs (applicativeApply (HsCon (UnQual constructor)) arguments)) []
+        bindingStatements <- foldMapM bindingStatement $
+          zip3 fieldNamesEtc qualifiedFieldNames constructors
 
-            [stmt0, stmt1] ++ inferenceStatement
+        let returnStatement = HsQualifier (HsApp returnE (HsParen namedSchema))
 
+        pure $ HsDo (bindingStatements ++ [ returnStatement ])
 
-          returnStatement = HsQualifier (HsApp returnE (HsParen namedSchema))
+  expression <- case maybeConstructors of
+    Nothing           -> expressionForMessage
+    Just constructors -> expressionForOneOf constructors
 
   let instanceDeclaration =
           instDecl_ className [ classArgument ] [ classDeclaration ]
@@ -1329,10 +1422,6 @@ toSchemaInstanceDeclaration messageName maybeConstructors fieldNames = do
             where
               match = match_ matchName [ HsPWildCard ] rightHandSide []
                 where
-                  expression = case maybeConstructors of
-                      Nothing           -> expressionForMessage
-                      Just constructors -> expressionForOneOf constructors
-
                   rightHandSide = HsUnGuardedRhs expression
 
                   matchName = HsIdent "declareNamedSchema"
@@ -1671,24 +1760,24 @@ dotProtoFieldC, primC, repeatedC, nestedRepeatedC, namedC, mapC,
   convertServerReaderHandlerE, convertServerWriterHandlerE,
   convertServerRWHandlerE, clientRegisterMethodE, clientRequestE :: HsExp
 
-dotProtoFieldC       = HsVar (protobufName "DotProtoField")
-primC                = HsVar (protobufName "Prim")
-repeatedC            = HsVar (protobufName "Repeated")
-nestedRepeatedC      = HsVar (protobufName "NestedRepeated")
-namedC               = HsVar (protobufName "Named")
-mapC                 = HsVar (protobufName "Map")
+dotProtoFieldC       = HsVar (protobufASTName "DotProtoField")
+primC                = HsVar (protobufASTName "Prim")
+repeatedC            = HsVar (protobufASTName "Repeated")
+nestedRepeatedC      = HsVar (protobufASTName "NestedRepeated")
+namedC               = HsVar (protobufASTName "Named")
+mapC                 = HsVar (protobufASTName "Map")
 fieldNumberC         = HsVar (protobufName "FieldNumber")
-singleC              = HsVar (protobufName "Single")
-pathC                = HsVar (protobufName "Path")
-dotsC                = HsVar (protobufName "Dots")
+singleC              = HsVar (protobufASTName "Single")
+pathC                = HsVar (protobufASTName "Path")
+dotsC                = HsVar (protobufASTName "Dots")
 nestedC              = HsVar (protobufName "Nested")
-anonymousC           = HsVar (protobufName "Anonymous")
-dotProtoOptionC      = HsVar (protobufName "DotProtoOption")
-identifierC          = HsVar (protobufName "Identifier")
-stringLitC           = HsVar (protobufName "StringLit")
-intLitC              = HsVar (protobufName "IntLit")
-floatLitC            = HsVar (protobufName "FloatLit")
-boolLitC             = HsVar (protobufName "BoolLit")
+anonymousC           = HsVar (protobufASTName "Anonymous")
+dotProtoOptionC      = HsVar (protobufASTName "DotProtoOption")
+identifierC          = HsVar (protobufASTName "Identifier")
+stringLitC           = HsVar (protobufASTName "StringLit")
+intLitC              = HsVar (protobufASTName "IntLit")
+floatLitC            = HsVar (protobufASTName "FloatLit")
+boolLitC             = HsVar (protobufASTName "BoolLit")
 forceEmitC           = HsVar (protobufName "ForceEmit")
 encodeMessageFieldE  = HsVar (protobufName "encodeMessageField")
 decodeMessageFieldE  = HsVar (protobufName "decodeMessageField")
@@ -1805,7 +1894,7 @@ dpTypeE (Map k v)          = apply mapC            [ dpPrimTypeE k, dpPrimTypeE 
 -- | Translate a dot proto primitive type to a Haskell AST primitive type.
 dpPrimTypeE :: DotProtoPrimType -> HsExp
 dpPrimTypeE ty =
-    let wrap = HsVar . protobufName in
+    let wrap = HsVar . protobufASTName in
     case ty of
         Named n  -> apply namedC [ dpIdentE n ]
         Int32    -> wrap "Int32"
@@ -1836,11 +1925,12 @@ defaultImports ImportCustomisation{ icUsesGrpc, icStringType = StringType string
 #ifdef DHALL
     , importDecl_ (m "Proto3.Suite.DhallPB")  & qualified (m hsDhallPB) & everything
 #endif
-    , importDecl_ (m "Proto3.Suite.DotProto") & qualified protobufNS & everything
+    , importDecl_ (m "Proto3.Suite.DotProto") & qualified protobufASTNS & everything
     , importDecl_ (m "Proto3.Suite.JSONPB")   & qualified jsonpbNS   & everything
     , importDecl_ (m "Proto3.Suite.JSONPB")   & unqualified          & selecting  [s".=", s".:"]
     , importDecl_ (m "Proto3.Suite.Types")    & qualified protobufNS & everything
     , importDecl_ (m "Proto3.Wire")           & qualified protobufNS & everything
+    , importDecl_ (m "Proto3.Wire.Decode")    & qualified protobufNS & selecting  [i"Parser", i"RawField"]
     , importDecl_ (m "Control.Applicative")   & qualified haskellNS  & everything
     , importDecl_ (m "Control.Applicative")   & unqualified          & selecting  [s"<*>", s"<|>", s"<$>"]
     , importDecl_ (m "Control.DeepSeq")       & qualified haskellNS  & everything
@@ -1875,6 +1965,7 @@ defaultImports ImportCustomisation{ icUsesGrpc, icStringType = StringType string
     grpcNS                    = m "HsGRPC"
     jsonpbNS                  = m "HsJSONPB"
     protobufNS                = m "HsProtobuf"
+    protobufASTNS             = m "HsProtobufAST"
     proxyNS                   = m "Proxy"
 
     -- staged constructors for importDecl
@@ -1915,7 +2006,14 @@ defaultServiceDeriving = map haskellName [ "Generic" ]
 --
 
 apply :: HsExp -> [HsExp] -> HsExp
-apply f = HsParen . foldl HsApp f
+apply f = paren . foldl HsApp f
+
+applyMaybe :: Maybe HsExp -> HsExp -> HsExp
+applyMaybe = maybe id (\f x -> paren (HsApp f (paren x)))
+
+paren :: HsExp -> HsExp
+paren e@(HsParen _) = e
+paren e = HsParen e
 
 applicativeApply :: HsExp -> [HsExp] -> HsExp
 applicativeApply f = foldl snoc nil
@@ -1957,11 +2055,16 @@ unqual_ = UnQual . HsIdent
 uvar_ :: String -> HsExp
 uvar_ = HsVar . unqual_
 
-protobufType_, primType_, protobufWrapperType_ :: String -> HsType
+protobufType_, primType_, protobufStringType_, protobufBytesType_ :: String -> HsType
 protobufType_ = HsTyCon . protobufName
 primType_ = HsTyCon . haskellName
-protobufWrapperType_ =
-  HsTyApp (HsTyCon (protobufName "Wrapped")) . HsTyCon .  haskellName
+protobufStringType_ = HsTyApp (protobufType_ "String") . HsTyCon . haskellName
+protobufBytesType_ = HsTyApp (protobufType_ "Bytes") . HsTyCon . haskellName
+
+protobufFixedType_, protobufSignedType_, protobufWrappedType_ :: HsType -> HsType
+protobufFixedType_ = HsTyApp (protobufType_ "Fixed")
+protobufSignedType_ = HsTyApp (protobufType_ "Signed")
+protobufWrappedType_ = HsTyApp (HsTyCon (protobufName "Wrapped"))
 
 type_ :: String -> HsType
 type_ = HsTyCon . unqual_