+++ tests/TestCodeGen.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE LambdaCase          #-}
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -38,11 +39,15 @@ codeGenTests = testGroup "Code generator unit tests"
   , camelCaseMessageFieldNames
   , don'tAlterEnumFieldNames
   , knownTypeMessages
-  , simpleEncodeDotProto "Binary"
-  , simpleDecodeDotProto "Binary"
-  , simpleEncodeDotProto "Jsonpb"
-  , simpleDecodeDotProto "Jsonpb"
+  , pythonInteroperation
   ]
+
+pythonInteroperation :: TestTree
+pythonInteroperation = testGroup "Python interoperation" $ do
+  tt <- ["Data.Text.Lazy.Text", "Data.Text.Text", "Data.Text.Short.ShortText"]
+  format <- ["Binary", "Jsonpb"]
+  direction <- [simpleEncodeDotProto, simpleDecodeDotProto]
+  [direction tt format]
 
 knownTypeMessages :: TestTree
 knownTypeMessages =
@@ -104,13 +109,21 @@ setPythonPath :: IO ()
 setPythonPath = Turtle.export "PYTHONPATH" .
   maybe pyTmpDir (\p -> pyTmpDir <> ":" <> p) =<< Turtle.need "PYTHONPATH"
 
-simpleEncodeDotProto :: T.Text -> TestTree
-simpleEncodeDotProto format =
-    testCase ("generate code for a simple .proto and then use it to encode messages in format " ++ show format)
+simpleEncodeDotProto :: String -> T.Text -> TestTree
+simpleEncodeDotProto chosenStringType format =
+    testCase
+    ("generate code for a simple .proto and then use it to encode messages" ++
+     " with string type " ++ chosenStringType ++ " in format " ++ show format)
     $ do
-         compileTestDotProtos
+         decodedStringType <- either die pure (parseStringType chosenStringType)
+
+         compileTestDotProtos decodedStringType
          -- Compile our generated encoder
-         Turtle.proc "tests/encode.sh" [hsTmpDir] empty >>= (@?= ExitSuccess)
+         let args = [hsTmpDir]
+#if DHALL
+                    ++ ["-DDHALL"]
+#endif
+         Turtle.proc "tests/encode.sh" args empty >>= (@?= ExitSuccess)
 
          -- The python encoder test exits with a special error code to indicate
          -- all tests were successful
@@ -122,13 +135,21 @@ simpleEncodeDotProto format =
          Turtle.rmtree hsTmpDir
          Turtle.rmtree pyTmpDir
 
-simpleDecodeDotProto :: T.Text -> TestTree
-simpleDecodeDotProto format =
-    testCase ("generate code for a simple .proto and then use it to decode messages in format " ++ show format)
+simpleDecodeDotProto :: String -> T.Text -> TestTree
+simpleDecodeDotProto chosenStringType format =
+    testCase
+    ("generate code for a simple .proto and then use it to decode messages" ++
+     " with string type " ++ chosenStringType ++ " in format " ++ show format)
     $ do
-         compileTestDotProtos
+         decodedStringType <- either die pure (parseStringType chosenStringType)
+
+         compileTestDotProtos decodedStringType
          -- Compile our generated decoder
-         Turtle.proc "tests/decode.sh" [hsTmpDir] empty >>= (@?= ExitSuccess)
+         let args = [hsTmpDir]
+#if DHALL
+                    ++ ["-DDHALL"]
+#endif
+         Turtle.proc "tests/decode.sh" args empty >>= (@?= ExitSuccess)
 
          setPythonPath
          let cmd = "python tests/send_simple_dot_proto.py " <> format <> " | FORMAT=" <> format <> " " <> hsTmpDir <> "/simpleDecodeDotProto "
@@ -144,17 +165,17 @@ simpleDecodeDotProto format =
 dumpAST :: [FilePath] -> FilePath -> IO ()
 dumpAST incs fp = either (error . show) putStrLn <=< runExceptT $ do
   (dp, tc) <- readDotProtoWithContext incs fp
-  renderHsModuleForDotProto theStringType mempty dp tc
+  renderHsModuleForDotProto defaultStringType mempty dp tc
 
 hsTmpDir, pyTmpDir :: IsString a => a
 hsTmpDir = "test-files/hs-tmp"
 pyTmpDir = "test-files/py-tmp"
 
-theStringType :: StringType
-theStringType = StringType "Data.Text.Lazy" "Text"
+defaultStringType :: StringType
+defaultStringType = StringType "Data.Text.Lazy" "Text"
 
-compileTestDotProtos :: IO ()
-compileTestDotProtos = do
+compileTestDotProtos :: StringType -> IO ()
+compileTestDotProtos decodedStringType = do
   Turtle.mktree hsTmpDir
   Turtle.mktree pyTmpDir
   let protoFiles =
@@ -173,10 +194,10 @@ compileTestDotProtos = do
   forM_ protoFiles $ \protoFile -> do
     compileDotProtoFileOrDie
         CompileArgs{ includeDir = ["test-files"]
-                   , extraInstanceFiles = []
+                   , extraInstanceFiles = ["test-files" Turtle.</> "Orphan.hs"]
                    , outputDir = hsTmpDir
                    , inputProto = protoFile
-                   , stringType = theStringType
+                   , stringType = decodedStringType
                    }
 
     let cmd = T.concat [ "protoc --python_out="