+++ tests/decode.sh
@@ -1,16 +1,19 @@
 set -eu
 hsTmpDir=$1
+shift
 
 ghc                                         \
     --make                                  \
     -odir $hsTmpDir                         \
     -hidir $hsTmpDir                        \
     -o $hsTmpDir/simpleDecodeDotProto       \
+    "$@"                                    \
     $hsTmpDir/TestProto.hs                  \
     $hsTmpDir/TestProtoImport.hs            \
     $hsTmpDir/TestProtoOneof.hs             \
     $hsTmpDir/TestProtoOneofImport.hs       \
     $hsTmpDir/TestProtoWrappers.hs          \
+    tests/Test/Dhall/Orphan.hs              \
     tests/SimpleDecodeDotProto.hs           \
     >/dev/null